refactor: move engine related classes
