+++ eze/src/main/kotlin/org/camunda/community/eze/engine/ZeebeEngineClockImpl.kt
@@ -5,9 +5,10 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package org.camunda.community.eze
+package org.camunda.community.eze.engine
 
 import io.camunda.zeebe.scheduler.clock.ControlledActorClock
+import org.camunda.community.eze.ZeebeEngineClock
 import java.time.Duration
 import java.time.Instant
